create endpoint crud team member of a project

+++ src/controllers/projectController.ts
@@ -1,12 +1,16 @@
 import type { Request, Response } from "express";
 import Project from "../models/Project";
+import Task from "../models/Task";
 
 export class ProjectController {
+
     static createProject = async (req: Request, res: Response) => {
-        console.log(req.user);
         
+        const project = new Project(req.body)
+        project.manager = req.user.id
+
         try {
-            await Project.create(req.body)
+            await project.save()
             res.json({message:"Proyecto creado Correctamente"})
             
         } catch (error) {
@@ -17,7 +21,7 @@ export class ProjectController {
     static getAllProjects = async (req: Request, res: Response) => {
 
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({manager: req.user.id})
             res.json(projects)
         } catch (error) {
             console.log(error);
@@ -32,9 +36,16 @@ export class ProjectController {
 
         try {
             const project = await Project.findById(id).populate("tasks")
+            
             if (!project) {
                 const error = new Error("Proyecto no encontrado")
                 res.status(404).json({error : error.message});
+                return
+            }
+            
+            if(project.manager!.toString() !== req.user.id.toString()) {
+                const error = new Error("No tienes permisos para ver este proyecto")
+                res.status(403).json({error : error.message});
                 return
             }
 
@@ -57,6 +68,13 @@ export class ProjectController {
                 res.status(404).json({error : error.message});
                 return
             }
+
+            if(project.manager!.toString() !== req.user.id.toString()) {
+                const error = new Error("No tienes permisos para modificar este proyecto")
+                res.status(403).json({error : error.message});
+                return
+            }
+
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -82,6 +100,14 @@ export class ProjectController {
                 return
             }
 
+            if(project.manager!.toString() !== req.user.id.toString()) {
+                const error = new Error("No tienes permisos para eliminar este proyecto")
+                res.status(403).json({error : error.message});
+                return
+            }
+
+            await Task.deleteMany({project: id})
+                
             await project.deleteOne()
             res.json({message:"Proyecto Eliminado"})
